+++ gamemodes/terrortown/gamemode/server/sv_main.lua
@@ -171,6 +171,14 @@ local ttt_minply = CreateConVar("ttt_minimum_players", "2", {FCVAR_NOTIFY, FCVAR
 
 ---
 -- @realm server
+local cvPreferMapModels = CreateConVar("ttt2_prefer_map_models", "1", {FCVAR_NOTIFY, FCVAR_ARCHIVE})
+
+---
+-- @realm server
+local cvSelectModelPerRound = CreateConVar("ttt2_select_model_per_round", "1", {FCVAR_NOTIFY, FCVAR_ARCHIVE})
+
+---
+-- @realm server
 CreateConVar("ttt2_prep_respawn", "0", {FCVAR_NOTIFY, FCVAR_ARCHIVE}, "Respawn if dead in preparing time")
 
 ---
@@ -311,6 +319,9 @@ function GM:Initialize()
 	ShopEditor.SetupShopEditorCVars()
 	ShopEditor.CreateShopDBs()
 
+	-- initialize playermodel database
+	playermodels.InitializeDatabase()
+
 	-- Force friendly fire to be enabled. If it is off, we do not get lag compensation.
 	RunConsoleCommand("mp_friendlyfire", "1")
 
@@ -337,7 +348,7 @@ function GM:Initialize()
 
 	self.DamageLog = {}
 	self.LastRole = {}
-	self.playermodel = GetRandomPlayerModel()
+	self.playermodel = playermodels.GetRandomPlayerModel()
 	self.playercolor = COLOR_WHITE
 
 	-- Delay reading of cvars until config has definitely loaded
@@ -508,6 +519,9 @@ function GM:InitPostEntity()
 	TTT2ShopFallbackInitialized = true
 
 	WEPS.ForcePrecache()
+
+	-- precache player models
+	playermodels.PrecacheModels()
 
 	timer.Simple(0, function()
 		addonChecker.Check()
@@ -1041,8 +1055,12 @@ function PrepareRound()
 	-- Update damage scaling
 	KARMA.RoundPrepare()
 
-	-- New look. Random if no forced model set.
-	GAMEMODE.playermodel = GAMEMODE.force_plymodel == "" and GetRandomPlayerModel() or GAMEMODE.force_plymodel
+	-- New look. Random if no forced model set
+	if cvPreferMapModels:GetBool() and GAMEMODE.force_plymodel and GAMEMODE.force_plymodel ~= "" then
+		GAMEMODE.playermodel = GAMEMODE.force_plymodel
+	elseif cvSelectModelPerRound:GetBool() then
+		GAMEMODE.playermodel = playermodels.GetRandomPlayerModel()
+	end
 
 	---
 	-- @realm server